Fix some issues with Logical Pickup Placement
Move the setup of the victory condition for logical pickup placement to the
logic bootrap. This means both the generator and resolver will both use
the modified citory condition.

Also include pickups configured to use their vanilla locations.
